docs: more doc comments
