+++ src/utils.ts
@@ -49,7 +49,7 @@ export async function newAccountWithLamports(
     if (--retries <= 0) {
       break;
     }
-    console.log('airdrop retry ' + retries);
+    // console.log('airdrop retry ' + retries);
   }
   throw new Error(`airdrop of ${lamports} failed`);
 }
